+++ FuzzySharp/SimilarityRatio/Scorer/Generic/ScorerBase.cs
@@ -4,5 +4,5 @@ namespace Raffinert.FuzzySharp.SimilarityRatio.Scorer.Generic;
 
 public abstract class ScorerBase<T> : IRatioScorer<T> where T : IEquatable<T>
 {
-    public abstract int Score(T[] input1, T[] input2);
+    public abstract double Score(T[] input1, T[] input2);
 }
